Create the endpoint for user sign-in
In this commit we create the POST signin endpoint,
so that the user can sign in with his email and
password.

The server returns a jwt token in the response and
also creates a cookie with that token.

We also add a sing-in validator to ensure that we
accept the data only in the appropriate format.