Create authorization for users
In this commit we create the logic in order to
apply authorization for a specific user to be
able to do certain actions.

The user is authorized when he sends his user id
as a parameter and his jwt token (with the same
user id).

We create two testing endpoints, one to test the
authentication and one to test the authorization
of the user.