Create API documentation in Swagger
In this commit we add the initial documentation
for the following endpoints:
- Signup
- Signin
- Signout
- Test

This documentation (in the form of an OpenAPI 3.0 document)
is available via the Swagger UI, in localhost:8080/api.

This was accomplished with the Swagger UI Express
module (https://www.npmjs.com/package/swagger-ui-express).

The documentation could be used later to validate requests
based on the OpenAPI document created, with the help of the
express-openapi-validate module
(https://www.npmjs.com/package/express-openapi-validate).